move around and clean up msw handlers
