+++ src/components/Orcid/__tests__/WorksTable.test.tsx
@@ -1,11 +1,12 @@
 import { expect, test, TestContext } from 'vitest';
-import { apiHandlerRoute, createServerListenerMocks, render, urls } from '@test-utils';
+import { createServerListenerMocks, render, urls } from '@test-utils';
 import { WorksTable } from '@components/Orcid';
 import { rest } from 'msw';
 import { ApiTargets, IADSApiSearchResponse } from '@api';
 import { equals } from 'ramda';
 import { waitFor } from '@testing-library/dom';
 import orcidProfileResponse from '@mocks/responses/orcid/orcid-profile_full.json';
+import { apiHandlerRoute } from '@mocks/mockHelpers';
 
 test.skip('triggers call to profile on mount', async ({ server }: TestContext) => {
   const { onRequest } = createServerListenerMocks(server);
