+++ src/mocks/handlers/objects.ts
@@ -1,7 +1,7 @@
 import { ApiTargets } from '@api';
 import { rest } from 'msw';
+import { apiHandlerRoute } from '@mocks/mockHelpers';
 
-const apiHandlerRoute = (key: ApiTargets, path?: string) => `*${key}${typeof path === 'string' ? path : '*'}`;
 export const objectsHandlers = [
   // on post to objects service, just return the query
   rest.post(apiHandlerRoute(ApiTargets.SERVICE_OBJECTS_QUERY), async (req, res, ctx) => {
